{{ cookiecutter.project_slug }}/requirements/dev.txt
black==19.10b0
pre-commit==2.7.1

# Debugging
django-debug-toolbar==2.2
